Motorbike fixes

+++ TombEngine/Objects/TR4/Vehicles/motorbike_info.h
@@ -1,7 +1,7 @@
 #pragma once
 struct MotorbikeInfo {
-	int wheelRight;  // (two wheel: front and back)
-	int wheelLeft;   // (one wheel: left)
+	short wheelRight;  // (two wheel: front and back)
+	short wheelLeft;   // (one wheel: left)
 	int velocity;
 	int revs;
 	int engineRevs;
